ExtractAccessKey now stores position
git-svn-id: svn://ultimatepp.org/upp/trunk@997 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/CtrlLib/MenuImp.h
@@ -39,7 +39,7 @@ protected:
 	Font    font;
 	bool    isenabled;
 	byte    type;
-	byte    accesskey;
+	int     accesskey;
 	Size    maxiconsize;
 	const MenuBar::Style *style;
